language: Structure-pack most used types
Make use of C++11 typed-enums and C++11 unions

sizeof-statistics (64 bit system):
- DUChainBaseData: 24B -> 20B
- AbstractTypeData: 32B -> 12B
- DelayedTypeData: 40B -> 24B
- DeclarationData: 76B -> 60B
- DUContextData: 64B -> 56B
- TopDUContextData: 96B -> 80B (not instantiated *that* often)

Differential Revision: https://phabricator.kde.org/D322

+++ language/duchain/types/abstracttype.cpp
@@ -47,12 +47,12 @@ AbstractType::AbstractType( AbstractTypeData& dd )
 {
 }
 
-quint64 AbstractType::modifiers() const
+quint32 AbstractType::modifiers() const
 {
   return d_func()->m_modifiers;
 }
 
-void AbstractType::setModifiers(quint64 modifiers)
+void AbstractType::setModifiers(quint32 modifiers)
 {
   d_func_dynamic()->m_modifiers = modifiers;
 }
